Move pkg/{apiserver -> genericapiserver}/authenticator
